% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_Credit.R
\name{ES_Credit}
\alias{ES_Credit}
\title{Compute Expected Shortfall (ES) of a Credit Risk}
\usage{
ES_Credit(credit, pd, alpha)
}
\arguments{
\item{credit}{Credit that was taken or given out}

\item{pd}{Probability of default}

\item{alpha}{Confidence Niveau}
}
\value{
numeric value
}
\description{
Compute Expected Shortfall (ES) of a Credit Risk
}
