% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_p.R
\name{mu_p}
\alias{mu_p}
\title{Compute Portfolio Return}
\usage{
mu_p(weights, returns, verbose = TRUE)
}
\arguments{
\item{weights}{weights of portfolio positions}

\item{returns}{returns of the positions of the portfolio}

\item{verbose}{prints out the result additionally}
}
\value{
numeric value
}
\description{
Compute Portfolio Return
}
