% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_p.R
\name{sigma_p}
\alias{sigma_p}
\title{Compute Portfolio Variance and Standard Deviation (Volatility)}
\usage{
sigma_p(weights, correlation_matrix, volatility, verbose = TRUE)
}
\arguments{
\item{weights}{weights of portfolio positions}

\item{correlation_matrix}{correlation matrix of the portfolio}

\item{volatility}{volatility of single assets}

\item{verbose}{prints out the result additionally}
}
\description{
Compute Portfolio Variance and Standard Deviation (Volatility)
}
