% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_Credit.R
\name{VaR_Credit}
\alias{VaR_Credit}
\title{Compute Value at Risk (VaR) of a Credit Risk}
\usage{
VaR_Credit(credit, pd, alpha)
}
\arguments{
\item{credit}{Credit that was taken or given out}

\item{pd}{Probability of default}

\item{alpha}{Confidence Niveau}
}
\value{
numeric value
}
\description{
Compute Value at Risk (VaR) of a Credit Risk
}
